+++ packages/js/product-editor/src/components/iframe-editor/index.ts
@@ -1,2 +1,2 @@
 export * from './iframe-editor';
-export * from './plugin-sidebar';
+export * from './sidebar/plugin-sidebar';
