+++ packages/js/product-editor/src/components/iframe-editor/sidebar/plugin-sidebar/index.js
